fix bugs and warning of implicit declaration

+++ program/system/remote_control.c
@@ -7,7 +7,7 @@ vs16 PWM_M4 = PWM_MOTOR_MIN;
 
 static __IO uint16_t rc_ctrl_timeout = 0;
 
-void Update_RC_Control(int16_t *Roll, int16_t  *Pitch, int16_t  *Yaw, int16_t  *Thr, uint8_t *safety)
+void Update_RC_Control(int16_t *Roll, int16_t  *Pitch, int16_t  *Yaw, int16_t  *Thr, int16_t *safety)
 {
 	/*Get PWM3 Input capture to control trottle*/
 	if ((system.variable[PWM3_CCR].value > MIN_PWM_INPUT) &&
